fix `--unreferenced=auto` clearing new pending snapshots
closes #751
